moved add member page
