+++ src/pages/HomePage/pages/Members/pages/AddMember.tsx
@@ -1,48 +1,50 @@
-import { pictureInstance as axiosPic } from '@/axiosInstance';
+import { pictureInstance as axiosPic } from "@/axiosInstance";
 import { pictureType } from "@/utils/interfaces";
 import { useNavigate } from "react-router-dom";
 import useState from "react-usestateref";
-import ProfilePicture from "../../../../components/ProfilePicture";
-import { memberValues } from "../../../../utils/helperFunctions";
-import { baseUrl } from "../../../Authentication/utils/helpers";
-import MembersForm from "./Components/MembersForm";
-import { addNewMember } from "./utils/apiCalls";
+import ProfilePicture from "../../../../../components/ProfilePicture";
+import { memberValues } from "../../../../../utils/helperFunctions";
+import { baseUrl } from "../../../../Authentication/utils/helpers";
+import MembersForm from "../Components/MembersForm";
+import { addNewMember } from "../utils/apiCalls";
 import editIcon from "/assets/home/edit.svg";
 
-
 const AddMember = () => {
-  const [profilePic, setProfilePic] = useState<pictureType>({ picture: "", src: "" });
+  const [profilePic, setProfilePic] = useState<pictureType>({
+    picture: "",
+    src: "",
+  });
   const [loading, setLoading] = useState(false);
   const [userValue, setUserValue, userValueRef] = useState(memberValues);
   const navigate = useNavigate();
-  function changePic(pic:pictureType) {
+  function changePic(pic: pictureType) {
     setProfilePic(() => pic);
   }
-  const handleChange = (name:string, value:string|boolean) => {
+  const handleChange = (name: string, value: string | boolean) => {
     setUserValue((prev) => ({ ...prev, [name]: value }));
-  }
+  };
   const handleCancel = () => {
     setUserValue(memberValues);
-    setProfilePic({picture:"",src:""});
+    setProfilePic({ picture: "", src: "" });
     navigate("/home/members");
-  }
+  };
   async function handleSubmit() {
-
     setLoading(true);
     const data = new FormData();
     data.append("file", profilePic.picture);
     const endpoint = "/upload";
     const path = `${baseUrl}${endpoint}`;
     try {
-      const response:any = profilePic.picture && await axiosPic.post(path, data);
+      const response: any =
+        profilePic.picture && (await axiosPic.post(path, data));
       if (profilePic.picture && response.status === 200) {
         const link = response.data.result.link;
-        setUserValue(prev => ({ ...prev, photo: link }));
+        setUserValue((prev) => ({ ...prev, photo: link }));
       }
       setLoading(true);
-      setProfilePic({picture:"",src:""});
-      const res = await addNewMember(userValueRef.current)
-      if(res && res.status === 200) navigate("/home/members"); // sends data after picture link is received
+      setProfilePic({ picture: "", src: "" });
+      const res = await addNewMember(userValueRef.current);
+      if (res && res.status === 200) navigate("/home/members"); // sends data after picture link is received
     } catch (error) {
       console.log(error);
       setLoading(false);
@@ -68,12 +70,22 @@ const AddMember = () => {
             onChange={changePic}
             id={"profilePic"}
           />
-          <div className="text-sm text-[#8F95B2] mt-3">Image size must be less <br /> than 2mb, jpeg or png</div>
+          <div className="text-sm text-[#8F95B2] mt-3">
+            Image size must be less <br /> than 2mb, jpeg or png
+          </div>
         </section>
       </div>
-      <MembersForm user={userValue} edit={true} onChange={handleChange} onSubmit={handleSubmit} onCancel={handleCancel} disabled={!userValue.email || !userValue.first_name || loading} loading={loading} />
+      <MembersForm
+        user={userValue}
+        edit={true}
+        onChange={handleChange}
+        onSubmit={handleSubmit}
+        onCancel={handleCancel}
+        disabled={!userValue.email || !userValue.first_name || loading}
+        loading={loading}
+      />
     </section>
   );
-}
+};
 
 export default AddMember;
